statistics: move IColumnStatisticEval to separate file
